FIX server login component for Vercel build
